Visualization, see description for key
Column 1 is the # of seconds passed at the end of the turn. The 5 columns are the part (if any) the worker is working on. The final column is the string of completed parts.

Note that a worker who just finished a part appears to be idle for the turn, but he actually spent that turn finishing the part.